step2/tickets_management/requirements.txt
flask
pika
python-consul